+++ NEWS.rst
@@ -4,6 +4,7 @@ NEWS for system-image updater
 
 0.7 (2013-XX-XX)
 ================
+ * No reboot should be issued if there is no update available.  (LP: #1202915)
 
 0.6 (2013-07-15)
 ================